1.when stored, never store a pointer;2.re-solve the precision problem: using tf.float64

+++ utils.py
@@ -3,15 +3,15 @@ import pandas as pd
 import tensorflow as tf
 
 def model_inputs(benign_dim, z_dim, attack_remains_dim):
-    inputs_benign = tf.placeholder(tf.float32, (None, benign_dim), name='input_benign') 
-    inputs_z = tf.placeholder(tf.float32, (None, z_dim), name='input_z')
-    inputs_attack_remains = tf.placeholder(tf.float32, (None, attack_remains_dim), name='input_attack_remains')
+    inputs_benign = tf.placeholder(tf.float64, (None, benign_dim), name='input_benign') 
+    inputs_z = tf.placeholder(tf.float64, (None, z_dim), name='input_z')
+    inputs_attack_remains = tf.placeholder(tf.float64, (None, attack_remains_dim), name='input_attack_remains')
     
     return inputs_benign, inputs_z, inputs_attack_remains
 
 def get_flow_dataset(filename=None):
     if filename is None:
-        feature = np.random.uniform(size=[20000,40]).astype(np.float32)
+        feature = np.random.uniform(size=[20000,40])
         label = np.random.randint(2, size=[20000,1])
         return np.concatenate([feature,label], axis=1)
     else:
